pyproject.toml
[project]
name = "rag-ai-webx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.125.0",
    "inngest>=0.5.13",
    "llama-index-core>=0.12.52.post1",
    "llama-index-readers-file>=0.4.8",
    "numpy>=2.3.5",
    "openai>=2.13.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "uvicorn>=0.38.0",
]
